fixed file include

+++ src/Http/Controllers/RoleController.php
@@ -16,16 +16,17 @@ class RoleController extends Controller
     public function __construct()
     {
         $this->middleware('auth:admin');
-        $this->authorize('isSuperAdmin', Admin::class);
     }
 
     public function index()
     {
+        $this->authorize('isSuperAdmin', Admin::class);
         return Role::all();
     }
 
     public function store(Request $request)
     {
+        $this->authorize('isSuperAdmin', Admin::class);
         $request->validate(['name' => 'required']);
         Role::create($request->all());
         return response('created', Response::HTTP_CREATED);
@@ -33,6 +34,7 @@ class RoleController extends Controller
 
     public function update(Role $role, Request $request)
     {
+        $this->authorize('isSuperAdmin', Admin::class);
         $request->validate(['name' => 'required']);
         $role->update($request->all());
         return response('updated', Response::HTTP_ACCEPTED);
@@ -40,6 +42,7 @@ class RoleController extends Controller
 
     public function destroy(Role $role)
     {
+        $this->authorize('isSuperAdmin', Admin::class);
         $role->delete();
         return response('deleted', Response::HTTP_NO_CONTENT);
     }
